docker_example/video_streaming/ai/requirements.txt
numpy
opencv-python
simple-tensor
pillow
